refactor dist utils, pep8
