+++ pytorch_lightning/metrics/metric.py
@@ -10,8 +10,8 @@ import torch
 from torch import nn
 
 from pytorch_lightning.utilities.apply_func import apply_to_collection
-from pytorch_lightning.metrics.utils import _flatten, gather_all_tensors_if_available
-from pytorch_lightning.metrics.utils import dim_zero_cat, dim_zero_mean, dim_zero_sum
+from pytorch_lightning.utilities.distributed import gather_all_tensors_if_available
+from pytorch_lightning.metrics.utils import _flatten, dim_zero_cat, dim_zero_mean, dim_zero_sum
 
 
 class Metric(nn.Module, ABC):
@@ -181,8 +181,8 @@ class Metric(nn.Module, ABC):
 
             if (
                 self._to_sync
-                and torch.distributed.is_available()
-                and torch.distributed.is_initialized()
+                and torch.distributed.is_available()  # noqa: W503
+                and torch.distributed.is_initialized()  # noqa: W503
             ):
                 self._sync_dist()
